beatonma-django/requirements.txt
# Dev tools
black
isort

# Build
wheel

beautifulsoup4>=4.12
celery>=5.2
django-colorfield>=0.7
django-extensions>=3.2
django-taggit>=3.0
django-user-agents>=0.4
django-wm>=4.1.0-dev0
Django>=4.2
gunicorn>=20.1
markdown2>=2.4
mf2py>=1.1
psycopg2-binary>=2.9
Pygments>=2.13
python-dateutil>=2.8
requests>=2.28
sentry-sdk>=1.18
Werkzeug>=2.2

pytest
pytest-django

# Handled by dockerfile
#git+ssh://git@github.com/beatonma/bmanotify.git
#git+ssh://git@github.com/beatonma/bmanotify-django.git
